rearranged codebase a bit, new login features to match google new scheme
