+++ bin/asterisk
@@ -1,4 +1,7 @@
 #!/usr/bin/env python
+"""
+Jacob Feisley, Ward Mundy, and Justin Quick
+"""
 from sys import exit
 from getpass import getpass
 from googlevoice.util import input, print_
