+++ bin/gvoice
@@ -2,10 +2,10 @@
 from sys import exit
 from atexit import register
 from optparse import OptionParser
-from googlevoice import Voice
+from googlevoice.voice import Voice
 from googlevoice.util import LoginError,pprint,input,print_
 
-parser = OptionParser(usage='''gvoice [-e <email> -p <password>] commands
+parser = OptionParser(usage='''gvoice [options] commands
     Where commands are
     
     login (li) - log into the voice service
@@ -38,6 +38,8 @@ parser.add_option("-p", "--password", dest='passwd', default=None,
                   help='Your account password (prompted if blank)')
 parser.add_option("-b", "--batch", dest='batch', default=False, action="store_true",
                   help='Batch operations, asking for no interactive input')
+parser.add_option("-d", "--debug", dest='debug', default=False, action="store_true",
+                  help='Print out debugging information')
 options, args = parser.parse_args()
 
 def login(email=options.email, passwd=options.passwd, batch=options.batch):
@@ -47,7 +49,7 @@ def login(email=options.email, passwd=options.passwd, batch=options.batch):
     global voice
     try:
         voice.login(options.email,options.passwd)
-    except LoginError:
+    except LoginError, e:
         if batch:
             print_('Login failed.')
             exit(0)
@@ -56,6 +58,11 @@ def login(email=options.email, passwd=options.passwd, batch=options.batch):
         else:
             exit(0)
 
+def logout():
+    global voice
+    print_('Logging out of voice...')
+    voice.logout()
+    
 def pprint_folder(name):
     folder = getattr(voice, name)()
     print_(folder)
@@ -70,10 +77,10 @@ if action == 'help':
     print_(parser.usage)
     exit(0)
 
-voice = Voice()
+voice = Voice(options.debug)
 login()
 
-register(voice.logout)
+register(logout)
 
 if action == 'interactive':
     while 1:
